ADD: relationship to table users

+++ database/migrations/2023_10_30_001639_create_bank_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->string('bank_name');
-            $table->integer('bank_number');
+            $table->bigInteger('bank_number');
             $table->timestamps();
         });
     }
